+++ spring-security/src/main/java/com/sap/cloud/security/spring/token/authentication/ReactiveHybridJwtDecoder.java
@@ -66,7 +66,7 @@ public class ReactiveHybridJwtDecoder implements ReactiveJwtDecoder {
     static Mono<Jwt> parseJwt(Token token) {
         try{
             Instant issuedAt = token.hasClaim(TokenClaims.XSUAA.ISSUED_AT)
-                    ? Instant.ofEpochSecond(Long.parseLong(token.getClaimAsString(TokenClaims.XSUAA.ISSUED_AT)))
+                    ? Instant.ofEpochSecond(Long.parseLong(token.getClaims().get(TokenClaims.XSUAA.ISSUED_AT).toString()))
                     : null;
             return Mono.just(new Jwt(token.getTokenValue(), issuedAt,
                     token.getExpiration(), token.getHeaders(), token.getClaims()));