+++ token-client/src/test/java/com/sap/cloud/security/xsuaa/util/HttpClientTestFactory.java
@@ -15,21 +15,36 @@ import org.apache.http.message.BasicHeader;
 import org.apache.http.message.BasicStatusLine;
 import org.mockito.Mockito;
 
+import java.util.Arrays;
+
+import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.when;
 
 public class HttpClientTestFactory {
 
 	public static CloseableHttpResponse createHttpResponse(String responseAsJson, int statusCode) {
+		Header[] headers = new Header[1];
+		headers[0] = new BasicHeader("testHeader", "testValue");
+		return createHttpResponse(responseAsJson, statusCode, headers);
+	}
+
+	public static CloseableHttpResponse createHttpResponse(String responseAsJson, int statusCode, Header[] headers) {
 		CloseableHttpResponse response = Mockito.mock(CloseableHttpResponse.class);
 		when(response.getStatusLine()).thenReturn(new BasicStatusLine(HttpVersion.HTTP_1_1, statusCode, null));
 		when(response.getEntity()).thenReturn(new StringEntity(responseAsJson, ContentType.APPLICATION_JSON));
-		Header[] headers = new Header[1];
-		headers[0] = new BasicHeader("testHeader", "testValue");
 		when(response.getAllHeaders()).thenReturn(headers);
+		Arrays.stream(headers).distinct().forEach(h -> {
+			lenient().when(response.containsHeader(h.getName())).thenReturn(true);
+			lenient().when(response.getFirstHeader(h.getName())).thenReturn(h);
+		});
 		return response;
 	}
 
 	public static CloseableHttpResponse createHttpResponse(String responseAsJson) {
 		return createHttpResponse(responseAsJson, HttpStatus.SC_OK);
+	}
+
+	public static CloseableHttpResponse createHttpResponseWithHeaders(String responseAsJson, Header[] headers) {
+		return createHttpResponse(responseAsJson, HttpStatus.SC_OK, headers);
 	}
 }